Enable metadata keys and values to be modified
This change introduces two new `Modifier.Target` instances: `metadataKeys`
and `metadataValues`, which can be used to customize how Ink parses metadata.
This in turn enables Publish plugins to be written that bridges the gap
between Publish’s built-in metadata format and that of external tools, such
as other static site generators or CMSs.